+++ packages/react/CHANGELOG.md
@@ -6,6 +6,10 @@ description: All notable changes to this project will be documented in this file
 
 ## [Unreleased]
 
+### Changed
+
+- Added `flushSync` to number input `onChange` handler to allow synchronous updates to the value when using the scrubber.
+
 ## [0.9.0] - 2023-07-21
 
 ### Added
@@ -146,5 +150,4 @@ description: All notable changes to this project will be documented in this file
 [0.7.3]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.7.3
 [0.8.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.8.0
 [0.8.1]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.8.1
-
 [0.9.0]: https://github.com/chakra-ui/ark/releases/tag/@ark-ui/react@0.9.0